fix: refactor project 2
