+++ 2_2d_animation/CMakeLists.txt
@@ -4,8 +4,10 @@ add_gl_executable_single_file(2_2d_animation src/main.cpp)
 target_sources(2_2d_animation PRIVATE
     src/TrailParticle.cpp
     src/Bullet.cpp
+    src/BulletPattern.cpp
     src/Player.cpp
     src/BossMove.cpp
+    src/MovePattern.cpp
     src/Boss.cpp
     src/HealthBar.cpp
     src/Background.cpp
